feat(api): 增强安全性、性能和 API 功能
安全改进:
- 敏感配置脱敏 (Redis URL、数据库凭据不返回给客户端)
- 添加认证中间件保护配置和 BizTag 端点
- 重构 SecureConfigResponse 结构

性能优化:
- AuthManager 使用 DashMap 替代 RwLock 实现无锁并发
- QpsWindow 使用 AtomicU64 计数器减少锁竞争

API 增强:
- list_biz_tags API 添加分页支持 (page/page_size/total)
- 添加 count_biz_tags 仓库方法
- 创建共享的 naive_to_rfc3339 工具函数

Bug 修复:
- 修复 SegmentInfo 可见性问题
- 移除冗余的类型转换
- 添加 #[allow(dead_code)] 到预留 API

+++ crates/core/src/database/connection.rs
@@ -51,8 +51,8 @@ pub async fn create_connection(config: &DatabaseConfig) -> Result<DatabaseConnec
 
     if config.engine != "sqlite" {
         connect_options
-            .max_connections(config.max_connections as u32)
-            .min_connections(config.min_connections as u32)
+            .max_connections(config.max_connections)
+            .min_connections(config.min_connections)
             .connect_timeout(std::time::Duration::from_secs(
                 config.acquire_timeout_seconds,
             ))
@@ -73,15 +73,18 @@ pub async fn create_connection(config: &DatabaseConfig) -> Result<DatabaseConnec
     Ok(db)
 }
 
+#[allow(dead_code)]
 pub async fn run_migrations(_db: &DatabaseConnection) -> Result<(), CoreError> {
     info!("Database migrations - migration system ready");
     Ok(())
 }
 
+#[allow(dead_code)]
 pub struct DatabaseManager {
     db: DatabaseConnection,
 }
 
+#[allow(dead_code)]
 impl DatabaseManager {
     pub fn new(db: DatabaseConnection) -> Self {
         Self { db }
